Rename tests for discovery and copy configs for isolation
