+++ app/unittests/test_sell_private_company_auction.py
@@ -11,8 +11,8 @@ from app.base import Move, MutableGameState
 from app.minigames.PrivateCompanyStockRoundAuction.minigame_auction import Auction
 from app.minigames.PrivateCompanyStockRoundAuction.minigame_decision import AuctionDecision
 from app.minigames.PrivateCompanyStockRoundAuction.move import AuctionDecisionMove, AuctionBidMove
-from app.unittests.PrivateCompanyMinigameTests import fake_player, fake_private_company
-from app.unittests.StockRoundMinigameTests import fake_public_company
+from app.unittests.test_private_company_minigame import fake_player, fake_private_company
+from app.unittests.test_stock_round_minigame import fake_public_company
 
 
 class AuctionRejectDecisionTests(unittest.TestCase):
